Make troi easily callable from python programs.
