+++ troi/cli.py
@@ -4,9 +4,7 @@ import sys
 import click
 import pytest
 
-import troi
-import troi.playlist
-import troi.utils
+from troi.core import generate_playlist, list_patches, patch_info
 
 
 @click.group()
@@ -47,105 +45,23 @@ def playlist(patch, debug, echo, save, token, upload, args, created_for, name, d
                     If it doesn't have sufficient numbers of tracks, ignore the playlist and don't submit it.
                     Default: Off, a playlist with at least one track will be considere complete.
     """
-
-    patchname = patch
-    patches = troi.utils.discover_patches()
-    if patch not in patches:
-        print("Cannot load patch '%s'. Use the list command to get a list of available patches." % patch,
-              file=sys.stderr)
-        sys.exit(1)
-
-    patch = patches[patch](debug)
-
-    context = patch.parse_args.make_context(patchname, list(args))
-    pipelineargs = context.forward(patch.parse_args)
-
-    patch_args = {
-        "echo": echo,
-        "save": save,
-        "token": token,
-        "created_for": created_for, 
-        "upload": upload, 
-        "name": name,
-        "desc": desc,
-        "min_recordings": min_recordings
-    }
-
-    pipeline = patch.create(pipelineargs, patch_args)
-    try:
-        playlist = troi.playlist.PlaylistElement()
-        playlist.set_sources(pipeline)
-        print("Troi playlist generation starting...")
-        result = playlist.generate()
-
-        if name:
-            playlist.playlists[0].name = name
-        if desc:
-            playlist.playlists[0].descripton = desc
-
-        print("done.")
-    except troi.PipelineError as err:
-        print("Failed to generate playlist: %s" % err, file=sys.stderr)
-        sys.exit(2)
-
-    if upload and not token:
-        print("In order to upload a playlist, you must provide an auth token. Use option --token.")
-        sys.exit(2)
-
-    if min_recordings is not None and \
-        (len(playlist.playlists) == 0 or len(playlist.playlists[0].recordings) < min_recordings):
-        print("Playlist does not have at least %d recordings, stopping." % min_recordings)
-        sys.exit(0)
-
-    if result is not None and token and upload:
-        for url, _ in playlist.submit(token, created_for):
-            print("Submitted playlist: %s" % url)
-
-    if result is not None and save:
-        playlist.save()
-        print("playlist saved.")
-
-    if result is not None and (echo or not token):
-        print()
-        playlist.print()
-
-    if not echo and not save and not token:
-        if result is None:
-            print("Patch executed successfully.")
-        elif len(playlist.playlists) == 0:
-            print("No playlists were generated. :(")
-        elif len(playlist.playlists) == 1:
-            print("A playlist with %d tracks was generated." % len(playlist.playlists[0].recordings))
-        else:
-            print("%d playlists were generated." % len(playlist.playlists))
-
-    sys.exit(0)
+    ret = generate_playlist(patch, debug, echo, save, token, upload, args, created_for, name, desc, min_recordings)
+    sys.exit(0 if ret else -1)
 
 
 @cli.command(name="list")
-def list_patches():
+def list_patches_cli():
     """List all available patches"""
-    patches = troi.utils.discover_patches()
-
-    print("Available patches:")
-    size = max([ len(k) for k in patches])
-    for slug in sorted(patches or []):
-        print("%s:%s %s" % (slug, " " * (size - len(slug)), patches[slug]().description()))
+    ret = list_patches()
+    sys.exit(0 if ret else -1)
 
 
 @cli.command()
 @click.argument("patch", nargs=1)
 def info(patch):
     """Get info for a given patch"""
-    patches = troi.utils.discover_patches()
-    if patch not in patches:
-        print("Cannot load patch '%s'. Use the list command to get a list of available patches." % patch,
-              file=sys.stderr)
-        sys.exit(1)
-
-    apatch = patches[patch]
-    context = click.Context(apatch.parse_args, info_name=patch)
-    click.echo(apatch.parse_args.get_help(context))
+    ret = patch_info(patch)
+    sys.exit(0 if ret else -1)
 
 
 @cli.command(context_settings=dict(
